[GDS2-172] [GDS2-174] Check the description for a huge paragraph
Did a bunch of testing & debugging for user story 172 as to what was causing some of the movement problems. Ive pinned the flying issue down to 3 things; 1.) The myth is getting spawned in the air in the first place... trying to fix that messes with a bunch of physics stuff, needs to be rectified. 2.) Some abilities have full on colliders which are interacting with the myth in ways they shouldnt. 3.) the sphere collider was causing weird issues where they would essentially roll on top of eachother, so i switched that out for a box collider

The movement issues (where players would slide with no input) seemed to be coming from the rigidbody component having its velocity / speed values updated, but not setting back to zero. The physics script we're using does not directly interact with the rigidbodies velocity. Manually setting this back to zero after it changes is a quick fix ive implemented, and havent been able to replicate the problem since (however more testing will come)

The other user story, 174, ive firstly removed swapping to hold. This is now going to be a process, whereby once its clicked, an animation will play and at the end of the animation it will swap. if the myth is hit during the animation, it gets cancelled. There will then be a cooldown on swapping again for a short amount of time. Ive added the state scripts for swapping & hooked it up, just need to now figure out the logic.